requirements.txt
langchain==0.3.9
langchain-core>=0.3.21
langchain-community==0.3.9
langchain-google-genai==2.0.6
langchain-cohere==0.3.3
langchain-groq==0.2.1
langchain-together==0.2.0
streamlit==1.40.2
PyGithub==2.1.1
python-dotenv==1.0.1
pyyaml==6.0.1
numpy==1.26.4
markdown==3.5.2
python-frontmatter==1.1.0
pandas==2.2.3